part_4 (Exercises 4.21, 4.22, 4.23)

+++ part4/Bloglist/controllers/blog.js
@@ -1,7 +1,7 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -10,27 +10,27 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-
   const token = request.token
   const decodedToken = jwt.verify(token, process.env.SECRET)
-
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   const user = await User.findById(decodedToken.id)
 
   if (!body.likes) {
     body.likes = 0
   }
+
   /*
   if (!blog.title && !blog.url) {
-    response
+    return response
       .status(400)
       .json({
         error: 'content missing',
       })
       .end()
-    return
   }
-  */
-
+*/
   const blog = new Blog({
     title: body.title,
     author: body.author,
@@ -56,22 +56,47 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if (blogToDelete.user._id.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).end()
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
   const body = request.body
-  const updateBlog = {
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+
+  const user = await User.findById(decodedToken.id)
+  const blogToUpdate = await Blog.findById(request.params.id)
+
+  if (blogToUpdate.user._id.toString() === user._id.toString()) {
+    const updateBlog = {
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+    }
+
+    const result = await Blog.findByIdAndUpdate(request.params.id, updateBlog, {
+      new: true,
+    })
+    response.json(result.toJSON())
+  } else {
+    return response.status(401).json({ error: `Unauthorized` })
   }
-  const result = await Blog.findByIdAndUpdate(request.params.id, updateBlog, {
-    new: true,
-  })
-  response.json(result)
 })
 
 module.exports = blogsRouter
